Fix a wrap-around issue with scratch space allocation.

+++ lib/nnc/mfa/ccv_nnc_mfa.cpp
@@ -196,7 +196,7 @@ MTL::Buffer* mfa::context::request_scratch(uint64_t size) {
   if (size > scratch->length()) {
     uint64_t padded_size = std::max(int64_t(0), int64_t(size) - 1);
     uint64_t leading_zeroes = __builtin_clzll(padded_size);
-    uint64_t rounded_size = 1 << uint64_t(64 - leading_zeroes);
+    uint64_t rounded_size = (uint64_t)1 << uint64_t(64 - leading_zeroes);
     
     auto buffer = device->newBuffer(rounded_size, MTL::ResourceStorageModePrivate | MTL::ResourceHazardTrackingModeTracked);
     CCV_NNC_MFA_PRECONDITION(buffer != nullptr);
